+++ src/main/java/io/github/gerritsmith/todobackspring/security/JwtUtilities.java
@@ -1,7 +1,6 @@
 package io.github.gerritsmith.todobackspring.security;
 
 import io.jsonwebtoken.*;
-import io.jsonwebtoken.security.SignatureException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
