+++ src/main/java/io/github/gerritsmith/todobackspring/user/api/LoginRequest.java
@@ -1,16 +1,11 @@
-package io.github.gerritsmith.todobackspring.user;
+package io.github.gerritsmith.todobackspring.user.api;
 
-import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 
-public class RegisterRequest {
+public class LoginRequest {
 
     @NotBlank
     private String username;
-
-    @NotBlank
-    @Email
-    private String email;
 
     @NotBlank
     private String password;
@@ -21,14 +16,6 @@ public class RegisterRequest {
 
     public void setUsername(String username) {
         this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
     }
 
     public String getPassword() {
