+++ src/main/java/io/github/gerritsmith/todobackspring/user/api/RegisterRequest.java
@@ -1,11 +1,16 @@
-package io.github.gerritsmith.todobackspring.user;
+package io.github.gerritsmith.todobackspring.user.api;
 
+import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 
-public class LoginRequest {
+public class RegisterRequest {
 
     @NotBlank
     private String username;
+
+    @NotBlank
+    @Email
+    private String email;
 
     @NotBlank
     private String password;
@@ -16,6 +21,14 @@ public class LoginRequest {
 
     public void setUsername(String username) {
         this.username = username;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
     }
 
     public String getPassword() {
