+++ src/main/java/io/github/gerritsmith/todobackspring/user/role/RoleType.java
@@ -1,4 +1,4 @@
-package io.github.gerritsmith.todobackspring.user;
+package io.github.gerritsmith.todobackspring.user.role;
 
 public enum RoleType {
     ROLE_USER,
